Figured it Out!
I was using the wrong function from previous. I needed to get the node
length which is a true reflection of the levels. I commented out the
section i didn’t need in the code.

Heads up!

This was a very helpful guide to piece a lot of this challenge together.
http://www.geeksforgeeks.org/level-order-tree-traversal/